mp_common: silence warning
int64_t was accidentally used with "%lld" format specifiers, which is
incorrect (even though long long int is always 64 bits, the type behind
int64_t can be different, e.g. it can be long int on 64 bit platforms).

+++ core/mp_common.c
@@ -31,7 +31,7 @@ char *mp_format_time(double time, bool fractions)
         sign[0] = '-';
     }
     long long int itime = time;
-    int64_t h, m, s;
+    long long int h, m, s;
     s = itime;
     h = s / 3600;
     s -= h * 3600;
